Moved Table.h into the platform directory, since it won't stay experimental. Made a proof of concept of the "index" -> class member offset map implemenation of the method for exposing public members to Lua.

+++ Platform/Platform.h
@@ -92,7 +92,7 @@ Tested in the field	:	NO
 
 #define BREAKPOINT(ID) __debugbreak();
 
-#define OFFSET_OF(CLASS,MEMBER) int( reinterpret_cast<long>( &(reinterpret_cast<CLASS*>(16)->MEMBER) ) - 16 )
+#define OFFSET_OF(CLASS,MEMBER) int( reinterpret_cast<unsigned long long>( &(reinterpret_cast<CLASS*>(16)->MEMBER) ) - 16 )
 
 
 enum BoolEnum
